# flatten recursive call to cat_shape