--- a/base/deprecated.jl
+++ b/base/deprecated.jl
@@ -230,4 +230,7 @@ function tuple_type_cons(::Type{S}, ::Type{T}) where T<:Tuple where S
     Tuple{S, T.parameters...}
 end
 
+# these were internal functions, but some packages seem to be relying on them
+@deprecate cat_shape(dims, shape::Tuple{}, shapes::Tuple...) cat_shape(dims, shapes)
+
 # END 1.6 deprecations