This brings a faster `cat` for arrays, especially for large ones

```julia
julia> s = [rand((0.0,1.0), 2,5,100) for _ = 1:5];

# before
julia> @btime cat(s...; dims=3);
  13.694 μs (83 allocations: 42.39 KiB)

# after
julia> @btime cat(s...; dims=3);
  10.816 μs (49 allocations: 41.19 KiB)

julia> s = [rand((0.0,1.0), 2,5,100) for _ = 1:5000];

# before
julia> @btime cat(s...; dims=3);
  751.367 ms (12552030 allocations: 803.21 MiB)

# after
julia> @btime cat(s...; dims=3);
  12.766 ms (30044 allocations: 40.10 MiB)
```

Similar results can be observed for `hcat` and `vcat`.

This also fixes #36750, would be nice to get this backported.

It is worth to note that there is actually a much slower workaround to #36750 

```julia
julia> s = [rand((0.0,1.0), 2,5,100) for _ = 1:5000];

julia> @btime reduce((x,y)->cat(x,y, dims=3), s);
  5.377 s (144963 allocations: 11.78 GiB)
```

and this is widely used in Flux community, e.g., https://github.com/FluxML/model-zoo/blob/4bf84760b370f264eadc7b19f6fb2070dd689bee/vision/dcgan_mnist/dcgan_mnist.jl#L27

cc: @cossio